FIX https://github.com/citizenos/citizenos-fe/issues/982

+++ src/app/topic/components/topic-member-user/topic-member-user.component.ts
@@ -9,6 +9,7 @@ import { ConfirmDialogComponent } from 'src/app/shared/components/confirm-dialog
 import { DialogService } from 'src/app/shared/dialog';
 import { Router } from '@angular/router';
 import { TranslateService } from '@ngx-translate/core';
+import { NotificationService } from 'src/app/services/notification.service';
 
 
 @Component({
@@ -27,6 +28,7 @@ export class TopicMemberUserComponent implements OnInit {
     private dialog: DialogService,
     private Translate: TranslateService,
     private TopicService: TopicService,
+    private Notification: NotificationService,
     private TopicMemberUserService: TopicMemberUserService,
     private router: Router,
   ) { }
@@ -71,6 +73,9 @@ export class TopicMemberUserComponent implements OnInit {
     });
     deleteUserDialog.afterClosed().subscribe(result => {
       if (result === true) {
+        if (!this.member.levelUser) {
+          return this.Notification.addError('COMPONENTS.TOPIC_MEMBER_USER.REMOVE_ERROR_MEMBER_VIA_GROUP');
+        }
         this.member.topicId = this.topic.id;
         this.TopicMemberUserService.delete({ topicId: this.topic.id, userId: this.member.userId || this.member.id })
           .pipe(take(1))
